Added brunch recipe for build

+++ bower/recipes/install.rb
@@ -4,7 +4,9 @@
 #
 node[:deploy].each do |application, deploy|
    execute "bower_install" do
+
         command "bower install --allow-root"
         cwd "#{deploy[:deploy_to]}/current"
+        only_if do :: File::exists?('./bower.json') end
    end
 end
